requirements.txt
Flask==3.1.0
google-generativeai==0.8.6
openpyxl==3.1.5
pandas==3.0.1
python-dotenv==1.2.2
